+++ docs/CHANGELOG.md
@@ -126,7 +126,8 @@
 
 ## 4.2.1-Dev 2025-09-08
 
-1. AlpacaFocuser.cpp/_getDeviceStateList_getDeviceStateList() TempCompAvailable and Connected added
+1. AlpacaFocuser.cpp/_getDeviceStateList() TempCompAvailable and Connected added
+2. AlpacaSwitch.cpp/_getDeviceStateList() SwitchName added, refactored